Refactor helper package
