+++ database/paginator.go
@@ -33,7 +33,7 @@ func paginateScope(page, pageSize int) func(db *gorm.DB) *gorm.DB {
 // filter results.
 //
 //  articles := []model.Article{}
-//  tx := database.Conn().Where("title LIKE ?", "%"+helper.EscapeLike(search)+"%")
+//  tx := database.Conn().Where("title LIKE ?", "%"+sqlutil.EscapeLike(search)+"%")
 //  paginator := database.NewPaginator(tx, page, pageSize, &articles)
 //  result := paginator.Find()
 //  if response.HandleDatabaseError(result) {
