--- helper/map.go
@@ -1,5 +1,0 @@
-package helper
-
-// Map is an alias to map[string]interface{}
-// Useful and a cleaner way to create a JSON response object
-type Map map[string]interface{}
