+++ util/httputil/httputil_bench_test.go
@@ -1,4 +1,4 @@
-package helper
+package httputil
 
 import (
 	"regexp"
