+++ util/httputil/sort.go
@@ -1,4 +1,4 @@
-package helper
+package httputil
 
 import "strings"
